+++ pajbot/utils.py
@@ -1,3 +1,4 @@
+import datetime
 import logging
 import subprocess
 import sys
@@ -76,3 +77,26 @@ def find(predicate, seq):
         if predicate(element):
             return element
     return None
+
+
+def json_serial(obj):
+    if isinstance(obj, datetime.datetime):
+        serial = obj.isoformat()
+        return serial
+    try:
+        return obj.jsonify()
+    except:
+        log.exception('Unable to serialize object with `jsonify`')
+        raise
+    raise TypeError('Type {} is not serializable'.format(type(obj)))
+
+
+def init_json_serializer(api):
+    import json
+    from flask import make_response
+
+    @api.representation('application/json')
+    def output_json(data, code, headers=None):
+        resp = make_response(json.dumps(data, default=json_serial), code)
+        resp.headers.extend(headers or {})
+        return resp