+++ pajbot/web/routes/base/stats.py
@@ -1,8 +1,10 @@
 from flask import render_template
 
 from pajbot.managers import DBManager
+from pajbot.managers import RedisManager
+from pajbot.managers import UserManager
 from pajbot.models.duel import UserDuelStats
-from pajbot.models.user import User
+from pajbot.streamhelper import StreamHelper
 
 
 def init(app):
@@ -10,9 +12,23 @@ def init(app):
     def stats():
         top_5_commands = sorted(app.bot_commands_list, key=lambda c: c.data.num_uses if c.data is not None else -1, reverse=True)[:5]
 
+        redis = RedisManager.get()
+
         # TODO: Make this hideable through some magic setting (NOT config.ini @_@)
         with DBManager.create_session_scope() as db_session:
-            top_5_line_farmers = db_session.query(User).order_by(User.num_lines.desc())[:5]
+            top_5_line_farmers = []
+            for redis_user in redis.zrevrangebyscore(
+                    '{streamer}:users:num_lines'.format(streamer=StreamHelper.get_streamer()),
+                    '+inf',
+                    '-inf',
+                    start=0,
+                    num=5,
+                    withscores=True,
+                    score_cast_func=int):
+                user = UserManager.get_static(redis_user[0], db_session=db_session)
+                user.save_to_redis = False
+                user.num_lines = redis_user[1]
+                top_5_line_farmers.append(user)
 
             return render_template('stats.html',
                     top_5_commands=top_5_commands,